refactor(topic-card): separate header from comment box
